64 build endpoints for investments and profit dist

+++ backend/internal/model/database/pitch.go
@@ -1,7 +1,7 @@
 package database
 
 type Pitch struct {
-	PitchID             *int    `json:"id,omitempty"`
+	PitchID             *int64  `json:"id,omitempty"`
 	ProductTitle        string  `json:"title"`
 	ElevatorPitch       string  `json:"elevator_pitch"`
 	DetailedPitch       string  `json:"detailed_pitch"`
@@ -10,5 +10,5 @@ type Pitch struct {
 	InvestmentStartDate string  `json:"investment_start_date"`
 	ProfitSharePercent  float64 `json:"profit_share_percent"`
 	UserID              string  `json:"user_id"`
-	RaisedAmount        int     `json:"raised_amount"`
+	RaisedAmount        int64   `json:"raised_amount"`
 }
